+++ samples/atomic/commerce-vite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @coveo/ui-kit-sample-atomic-commerce-vite
 
+## 3.60.5
+
+### Patch Changes
+
+- Updated dependencies [[`7fd1533`](https://github.com/coveo/ui-kit/commit/7fd15334c494ef03d11016fbf673d97cd9108e8e), [`64e9884`](https://github.com/coveo/ui-kit/commit/64e988411fce135d72a529d3569cd0657c75a1e8), [`48407c4`](https://github.com/coveo/ui-kit/commit/48407c44a7859ff42bf4195d4a65d495aa54565e), [`7ac71d6`](https://github.com/coveo/ui-kit/commit/7ac71d6f001239eaed0829048a5d6780b1484e21)]:
+  - @coveo/headless@3.55.0
+  - @coveo/atomic@3.60.5
+
 ## 3.60.4
 
 ### Patch Changes